Refactor layout directory structure
- Move container.tsx to ui/ (pure presentational primitive)
- Group action-bar files into layout/action-bar/ subdirectory
- Group footer files into layout/footer/ subdirectory
