+++ apps/template/app/layout.tsx
@@ -8,8 +8,8 @@ import { Suspense } from "react";
 import { AgentButton } from "@/components/agent/agent-button";
 import { CartProvider } from "@/components/cart/context";
 import { CartOverlayWithAddress } from "@/components/cart/overlay-with-address";
-import { ActionBar } from "@/components/layout/action-bar";
-import { Footer } from "@/components/layout/footer";
+import { ActionBar } from "@/components/layout/action-bar/action-bar";
+import { Footer } from "@/components/layout/footer/footer";
 import { Nav } from "@/components/layout/nav";
 import { SiteSchema } from "@/components/schema/site-schema";
 import { siteConfig } from "@/lib/config";
