+++ apps/template/app/not-found.tsx
@@ -1,7 +1,7 @@
 import { getTranslations } from "next-intl/server";
 import Link from "next/link";
 
-import { Container } from "@/components/layout/container";
+import { Container } from "@/components/ui/container";
 
 export default async function NotFoundError() {
   const t = await getTranslations("common");
