+++ apps/template/app/page.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { getTranslations } from "next-intl/server";
 
 import { BannerSection } from "@/components/banner-section";
-import { Container } from "@/components/layout/container";
+import { Container } from "@/components/ui/container";
 import { ProductsGrid } from "@/components/product/products-grid";
 import { siteConfig } from "@/lib/config";
 import { getLocale } from "@/lib/params";
