+++ apps/template/components/collections/collection-page.tsx
@@ -4,7 +4,7 @@ import { Suspense } from "react";
 
 import { CollectionResultsSection } from "@/components/collections/results";
 import { CollectionStructuredData } from "@/components/collections/structured-data";
-import { Container } from "@/components/layout/container";
+import { Container } from "@/components/ui/container";
 import { Skeleton } from "@/components/ui/skeleton";
 import { getCollectionResultsData, getCollectionSearchState } from "@/lib/collections/server";
 import type { Locale } from "@/lib/i18n";
