+++ apps/template/components/layout/action-bar/predictive-search-results.tsx
